fix:SpamCountとSpamTimeのやつをconfigからdataに移動
以上
